[Mozilla Branding Removal] Replace Mozilla Spoke feedback form link with a Discord invite
The Submit Feedback buttons in the Spoke project editor main menu (and its Help submenu) currently link to a Google form for sending feedback to Mozilla.  This commit replaces the Mozilla form links with a Discord invite link so feedback about Spoke will come to the Hubs Discord server instead.